xdiff: signed right shift for 32-bit case of XDL_FAST_HASH
In the 32-bit branch of count_masked_bytes(), the compiler uses a
logical right shift on mask, because it is unsigned.  We want it to
be an arithmetic right shift, however (keeping negative numbers
negative instead of shifting the sign bit as well).

There is no way to specify it in C, but we can at least cast mask to
signed long.  This will make the compiler use an arithmetic right shift
on certain implementations, but that's OK because we only turn on
XDL_FAST_HASH in the Makefile for known-good architectures, and at
least gcc 4.6.3 and clang 3.0 do what we want on the most interesting
one (x86).

+++ xdiff/xutils.c
@@ -311,7 +311,7 @@ static inline long count_masked_bytes(unsigned long mask)
 		 * (a+b+1) gives us
 		 *   correct 0-3 bytemask count result
 		 */
-		long a = (mask - 256) >> 23;
+		long a = ((long)mask - 256) >> 23;
 		long b = mask & 1;
 		return a + b + 1;
 	}
